Initial pre-processing step with imbalanced-learn
Making use of Synthetic Minory Oversampling Technique (SMOTE) from imbalanced-learn to account for the minority class where clients subscribe to a term deposit product. This accounts for ~12% of the dataset.
The bank-additional dataset files have been removed as the increased number of features and presence of missing values are not desired for this usecase.